requirements.txt
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-docx>=1.0.0
moviepy>=1.0.3
ebooklib>=0.18
PyPDF2>=3.0.1
pdfplumber>=0.9.0
volcengine-python-sdk[ark]>=1.0.94